feat: rituals
